libgit: add autogit support for LFS files
LFS files are stored in the git bare repo using a git object with
contents like this:

```
version https://git-lfs.github.com/spec/v1
oid sha256:0fa19837aa72b28c8045f997436d859ef177cda742cd7e7b87aa290c93b09e62
size 2300102
```

And then the actual file is stored under the `kbfs_lfs` directory
inside the bare git repo.  So in order to find out the size of the
file (to report when reading the autogit directory), we have to
actually read the file to find out if it's even an LFS file or not.
And it's much too slow to do that in autogit for every file in the
directory, because those block loads add up.

So let's play a trick.  It looks like the size of LFS git objects are
fairly fixed, varying only by the number of digits in the size of the
file.  So, we do this: if the repo does have an LFS subdirectory, and
the file in question has a size within a reasonable range for an LFS
file, only then do we check the contents of the file to see if it's
LFS.  If it is, we return a special file info type that identifies it
as LFS, and if it is, we use the OID to open the actual LFS file when
the file is opened.

Issue: HOTPOT-1305